SPYT-176 adding float type (#73)
* SPYT-176 added float and test

* SPYT-176 fixed bugs

* SPYT-176 debugging

* SPYT-176 fixed test

* SPYT-176 fixed codestyle

* SPYT-176 fixed test

* SPYT-176 removed schema hint in test

+++ taxi/dmp/spark/spark-over-yt/data-source/src/main/scala/ru/yandex/spark/yt/serializers/SchemaConverter.scala
@@ -89,7 +89,7 @@ object SchemaConverter {
     case IntegerType => YtLogicalType.Int32
     case LongType => YtLogicalType.Int64
     case StringType => YtLogicalType.String
-    case FloatType => YtLogicalType.Double
+    case FloatType => YtLogicalType.Float
     case DoubleType => YtLogicalType.Double
     case BooleanType => YtLogicalType.Boolean
     case _: ArrayType => YtLogicalType.Any
